requirements.txt
uvicorn >= 0.15.0
fastapi >= 0.68.1
logzero >= 1.7.0
python-dotenv >= 0.15.0
pydantic >= 1.7.3